Menu Program V1
This is about a simple menu system that is responsible for displaying menu buttons (new, display, and, exit) in addition to the current position and ASCII and name of every key pressed. This system functions as follows: clicking "new" will lead the user to enter (the name, age, ID, and, gender) of every user, and after all the user can exit or continue storing more students by choosing to click right-click or enter buttons. exit is not actually exit but it's supposed to 😂.